added index page to decks, modified pages/index to be a placeholder landing page
